# docs/swagger.yaml
components:
  schemas:
...
    StashTabMetadata:
      properties:
        colour:
          description: Colour 6 digit hex colour
          type: string
        folder:
          description: Folder always true if present
          type: boolean
        map:
          additionalProperties: {}
          description: Map various game specific properties
          type: object
        public:
          description: Public always true if present
          type: boolean
      type: object
...
    Action:
      enum:
      - added
      - modified
      - removed
      type: string
      x-enum-varnames:
      - ActionAdded
      - ActionModified
      - ActionRemoved
    AddGuildStashHistoryResponse:
      properties:
        number_of_added_entries:
          type: integer
      required:
      - number_of_added_entries
      type: object
...
    Event:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        teams:
          items:
            $ref: '#/components/schemas/Team'
          type: array
          uniqueItems: false
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - id
      - is_current
      - is_locked
      - is_main_event
      - is_public
      - max_size
      - name
      - teams
      - uses_medals
      - waitlist_size
      type: object
...
    Guild:
      properties:
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
        tag:
          type: string
        team_id:
          type: integer
      required:
      - id
      - name
      - tag
      type: object
    GuildStashChangeResponse:
      properties:
        entries:
          items:
            properties:
              account:
                properties:
                  name:
                    type: string
                type: object
              action:
                type: string
              id:
                type: string
              item:
                type: string
              league:
                type: string
              stash:
                type: string
              time:
                type: integer
              x:
                type: integer
              "y":
                type: integer
            type: object
          type: array
          uniqueItems: false
        truncated:
          type: boolean
      type: object
    GuildStashChangelog:
      properties:
        account_name:
          type: string
        action:
          $ref: '#/components/schemas/Action'
        item_name:
          type: string
        number:
          type: integer
        stash_name:
          type: string
        timestamp:
          type: integer
      required:
      - account_name
      - action
      - item_name
      - number
      - timestamp
      type: object
    GuildStashLogTimestampResponse:
      properties:
        earliest:
          type: integer
        latest:
          type: integer
        league_end:
          type: integer
        league_start:
          type: integer
      required:
      - league_end
      - league_start
      type: object
    GuildStashTab:
      properties:
        color:
          type: string
        fetch_enabled:
          type: boolean
        id:
          type: string
        index:
          type: integer
        last_fetch:
          format: date-time
          type: string
        name:
          type: string
        parent_id:
          type: string
        priority_fetch:
          type: boolean
        type:
          type: string
        user_ids:
          items:
            type: integer
          type: array
          uniqueItems: false
      required:
      - fetch_enabled
      - id
      - last_fetch
      - name
      - priority_fetch
      - type
      - user_ids
      type: object
...
    StashTabWithCompletions:
      properties:
        children:
          items:
            $ref: '#/components/schemas/StashTabWithCompletions'
          type: array
          uniqueItems: false
        folder:
          description: Folder a 10 digit hexadecimal string
          type: string
        id:
          description: Id a 10 digit hexadecimal string
          type: string
        index:
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ItemWithCompletions'
          type: array
          uniqueItems: false
        metadata:
          $ref: '#/components/schemas/StashTabMetadata'
        name:
          type: string
        parent:
          description: Parent a 10 digit hexadecimal string
          type: string
        type:
          type: string
      required:
      - id
      - metadata
      - name
      - type
      type: object
...
    TabSwitchRequest:
      properties:
        fetch_enabled:
          type: boolean
        priority_fetch:
          type: boolean
      type: object
    Team:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - event_id
      - id
      - name
      type: object
...
    GameVersion:
      enum:
      - poe1
      - poe2
      type: string
      x-enum-varnames:
      - PoE1
      - PoE2
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
  /{eventId}/guilds:
    get:
      description: Get all guilds for current event with their respective team ids
      operationId: GetGuilds
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Guild'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - guild-stash
  /{eventId}/guilds/{guildId}:
    put:
      description: Saves a guild for the current event
      operationId: SaveGuild
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Guild Id
        in: path
        name: guildId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Guild'
              description: Guild
              summary: guild
        description: Guild
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guild'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - guild-stash
  /{eventId}/guilds/{guildId}/stash-history:
    get:
      description: Fetches log entries for a guild in an event
      operationId: GetLogEntriesForGuild
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Guild Id
        in: path
        name: guildId
        required: true
        schema:
          type: integer
      - description: Limit
        in: query
        name: limit
        schema:
          type: integer
      - description: Offset
        in: query
        name: offset
        schema:
          type: integer
      - description: Name of the user doing the action (Make sure to replace the pound
          sign with a minus)
        in: query
        name: username
        schema:
          type: string
      - description: Name of the item (Can be partial)
        in: query
        name: itemname
        schema:
          type: string
      - description: Name of the stash tab
        in: query
        name: stashname
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GuildStashChangelog'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - guild-stash
    post:
      description: Adds a new entry to the guild stash history
      operationId: AddGuildstashHistory
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Guild Id
        in: path
        name: guildId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuildStashChangeResponse'
              description: Request body
              summary: guildStashChanges
        description: Request body
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddGuildStashHistoryResponse'
          description: Created
      security:
      - BearerAuth: []
      tags:
      - guild-stash
  /{eventId}/guilds/{guildId}/stash-history/latest_timestamp:
    get:
      description: Fetches the latest timestamp for a user's guild stash history
      operationId: GetLatestTimestampForUser
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Guild Id
        in: path
        name: guildId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildStashLogTimestampResponse'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - guild-stash
  /{eventId}/teams/{teamId}/guild-stash:
    get:
      description: Fetches all guild stash tabs for a user
      operationId: GetGuildStashForUser
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: teamId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GuildStashTab'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - guild-stash
  /{eventId}/teams/{teamId}/guild-stash/{stash_id}:
    get:
      description: Fetches a specific guild stash tab
      operationId: GetGuildStashTab
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: teamId
        required: true
        schema:
          type: integer
      - description: Stash Tab Id
        in: path
        name: stash_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StashTabWithCompletions'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - guild-stash
    patch:
      description: Enables fetching for a specific guild stash tab
      operationId: SwitchStashFetching
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: teamId
        required: true
        schema:
          type: integer
      - description: Stash Tab Id
        in: path
        name: stash_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TabSwitchRequest'
              description: Request body
              summary: body
        description: Request body
        required: true
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - guild-stash
  /{eventId}/teams/{teamId}/guild-stash/{stash_id}/update:
    post:
      description: Fetches current items for specific guild stash tab
      operationId: UpdateStashTab
      parameters:
      - description: Event Id
        in: path
        name: eventId
        required: true
        schema:
          type: integer
      - description: Team Id
        in: path
        name: teamId
        required: true
        schema:
          type: integer
      - description: Stash Tab Id
        in: path
        name: stash_id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - guild-stash
...
  /events/{event_id}:
    delete:
      description: Deletes an event
      operationId: DeleteEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - event
    get:
      description: Fetches an event by id
      operationId: GetEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...